+++ frontend/src/core/data/useTranslatedToolRegistry.tsx
@@ -39,6 +39,7 @@ import AutoRename from "@app/tools/AutoRename";
 import SingleLargePage from "@app/tools/SingleLargePage";
 import PageLayout from "@app/tools/PageLayout";
 import UnlockPdfForms from "@app/tools/UnlockPdfForms";
+import FormFill from "@app/tools/formFill/FormFill";
 import RemoveCertificateSign from "@app/tools/RemoveCertificateSign";
 import RemoveImage from "@app/tools/RemoveImage";
 import CertSign from "@app/tools/CertSign";
@@ -345,6 +346,19 @@ export function useTranslatedToolCatalog(): TranslatedToolCatalog {
         operationConfig: unlockPdfFormsOperationConfig,
         synonyms: getSynonyms(t, "unlockPDFForms"),
         automationSettings: null
+      },
+      formFill: {
+        icon: <LocalIcon icon="text-fields-rounded" width="1.5rem" height="1.5rem" />,
+        name: t('home.formFill.title', 'Fill Form'),
+        component: FormFill,
+        description: t('home.formFill.desc', 'Fill PDF form fields interactively with a visual editor'),
+        categoryId: ToolCategoryId.STANDARD_TOOLS,
+        subcategoryId: SubcategoryId.GENERAL,
+        workbench: 'viewer' as const,
+        endpoints: ['form-fill'],
+        automationSettings: null,
+        supportsAutomate: false,
+        synonyms: ['form', 'fill', 'fillable', 'input', 'field', 'acroform'],
       },
       changePermissions: {
         icon: <LocalIcon icon="lock-outline" width="1.5rem" height="1.5rem" />,
