+++ frontend/src/core/tools/formFill/formApi.ts
@@ -2,7 +2,7 @@
  * API service for form-related backend calls.
  */
 import apiClient from '@app/services/apiClient';
-import type { FormField } from '@proprietary/tools/formFill/types';
+import type { FormField } from '@app/tools/formFill/types';
 
 /**
  * Fetch form fields with coordinates from the backend.
