+++ frontend/src/core/types/toolId.ts
@@ -56,6 +56,7 @@ export const CORE_REGULAR_TOOL_IDS = [
   'showJS',
   'bookletImposition',
   'pdfTextEditor',
+  'formFill',
 ] as const;
 
 export const CORE_SUPER_TOOL_IDS = [
