--- frontend/src/proprietary/tools/formFill/providers/index.ts
@@ -1,3 +1,0 @@
-export type { IFormDataProvider } from '@proprietary/tools/formFill/providers/types';
-export { PdfLibFormProvider } from '@proprietary/tools/formFill/providers/PdfLibFormProvider';
-export { PdfBoxFormProvider } from '@proprietary/tools/formFill/providers/PdfBoxFormProvider';
