+++ frontend/src/proprietary/types/proprietaryToolId.ts
@@ -5,7 +5,6 @@
  */
 
 export const PROPRIETARY_REGULAR_TOOL_IDS = [
-  'formFill',
 ] as const;
 
 export const PROPRIETARY_SUPER_TOOL_IDS = [
